Cleanup mapselector and asset types
