+++ src/signals/incident/components/form/MapSelectors/Asset/Selector/WfsLayer/AssetLayer/AssetLayer.test.tsx
@@ -10,9 +10,9 @@ import type { MapOptions } from 'leaflet'
 import assetsJson from 'utils/__tests__/fixtures/assets.json'
 import MAP_OPTIONS from 'shared/services/configuration/map-options'
 
+import type { FeatureType } from 'signals/incident/components/form/MapSelectors/types'
 import { WfsDataProvider } from '../context'
 
-import type { FeatureType } from '../../../types'
 import type { ClusterMarker } from './AssetLayer'
 
 import { shouldSpiderfy, getMarkerByZoomLevel } from './AssetLayer'
