+++ src/signals/incident/components/form/MapSelectors/Asset/Selector/WfsLayer/WfsLayer.tsx
@@ -9,8 +9,8 @@ import type { FeatureCollection } from 'geojson'
 import type { Map as MapType } from 'leaflet'
 
 import AssetSelectContext from 'signals/incident/components/form/MapSelectors/Asset/context'
+import type { DataLayerProps } from 'signals/incident/components/form/MapSelectors/types'
 import useLayerVisible from '../../../hooks/useLayerVisible'
-import type { DataLayerProps } from '../../types'
 import { NO_DATA, WfsDataProvider } from './context'
 
 const SRS_NAME = 'urn:ogc:def:crs:EPSG::4326'
